Accept Google Maps API key as injected environment variable
Summary:
Pass Google Maps Javascript API key as environment variable rather than as a string hardcoded in source, in preparation for full open sourcing.
This is actually not super bad since this is a client-side API key, so it gets exposed to any client anyways when running the app. But this is a bit cleaner.
Old key has been destroyed.

Reviewers: #skynet

Differential Revision: https://code.skynet.engineering/D146